pyproject.toml
[tool.poetry]
name = "maestro"
version = "0.1.0"
description = "A multi-agent platform with the vision to facilitate deploy and run AI agents."
authors = ["IBM"]
license = "Apache 2.0"
readme = "README.md"
packages = [
    {include = "src", from="."},
    {include = "cli", from="."},
]

[tool.poetry.dependencies]
python = ">= 3.11, < 3.13"
openai = "^1.76.2"
pyyaml = "^6.0.2"
python-dotenv = "^1.0.1"
jsonschema = "^4.23.0"
docopt-ng = "^0.9.0"
langchain-community = "^0.3.16"
duckduckgo-search = "^7.3.0"
streamlit = "^1.44.0"
streamlit-mermaid = "^0.3.0"
psutil = "^7.0.0"
openapi = "^2.0.0"
openai-agents = {extras = ["litellm"], version = "^0.0.14"}
pycron = "^3.1.2"
beeai-framework = "^0.1.17"
slack_sdk = "^3.35.0"
nest-asyncio = "^1.6.0"
pydantic-ai = {extras = ["logfire"], version = "^0.1.8"}

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-v -s --ignore=framework"

[tool.poetry.scripts]
"maestro" = "cli.maestro:__run_cli"
